Updated MachineStates
The EnemyStates are no longer (EnemyStates) and they have become children classes of (CharacterStates), the reason for this is that NPCs share so many similar functionalities and the code felt so redundant.
