+++ src/elements/StationaryChaserEnemy.ts
@@ -1,7 +1,7 @@
 import { BodyHandle } from "engine/BodyHandle";
 import { StateMachineLMent, State } from "engine/StateMachineLMent";
 import { LookAt } from "./LookAt";
-import { EnemyAlertState, EnemyChaseState, EnemyIdleState, EnemyPatrolState, EnemyStates } from "./EnemyStates";
+import { CharacterStates, EnemyAlertState, EnemyChaseState, characterIdleState, characterPatrolState } from "./CharacterStates";
 
 
 export class StationaryChaserEnemy extends StateMachineLMent {
@@ -30,13 +30,13 @@ export class StationaryChaserEnemy extends StateMachineLMent {
         let point1 = this.body.body.getPosition().clone();
         let initQuat = this.body.body.getRotation().clone();
         this.states = {
-            [EnemyStates.patrol]: new EnemyPatrolState(this, [point1], this.backSpeed,this.alertZoneRadius,this.movementForce),
-            [EnemyStates.chase]: new EnemyChaseState(this, this.chaseSpeed, this.alertZoneRadius,this.movementForce),
-            [EnemyStates.alert]: new EnemyAlertState(this,this.alertZoneRadius,this.alertCooldown,this.alertWarmUp,EnemyStates.chase),
-            [EnemyStates.idle]: new EnemyIdleState(this,this.alertZoneRadius,999999,initQuat)
+            [CharacterStates.patrol]: new characterPatrolState(this, [point1], this.backSpeed,this.movementForce,this.alertZoneRadius),
+            [CharacterStates.chase]: new EnemyChaseState(this, this.chaseSpeed, this.alertZoneRadius,this.movementForce),
+            [CharacterStates.alert]: new EnemyAlertState(this,this.alertZoneRadius,this.alertCooldown,this.alertWarmUp,CharacterStates.chase),
+            [CharacterStates.idle]: new characterIdleState(this,this.alertZoneRadius,999999,undefined,initQuat)
         }
 
-        this.switchState(EnemyStates.patrol);
+        this.switchState(CharacterStates.patrol);
     }
 
     onStart() {
